refactor: add computer repository class and computer model class
